Creating server and db
So far I have begun creating the server as well as the database for the project. I am currently working on a function that will take the information from the form an place it into a localhost db on mongo. To run this code you'll need to install the npm dependencies as well as create a mongodb with port 27017 (which is default) and within this database have a Users database will a collection of users. Finally, to run this program within terminal (similar tool) go into the log-in dir, and type 'nodemon index'. this commend will automatically update the server on any changes.